% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_problems.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Knapsack Algorithm}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe with item weights and values.}

\item{W}{The maximum weight capacity.}
}
\value{
The best value that can be achieved within the weight limit.
}
\description{
Solves the knapsack problem using a greedy algorithm that selects items based on the highest value-to-weight ratio.
}
\examples{
items <- data.frame(weight=c(1,2,3), value=c(6,10,12))
W <- 5
greedy_knapsack(items, W)
}
