% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_problems.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack Dynamic Programming Solution}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A dataframe where each row is an item with two columns: weight and value.}

\item{W}{The total weight capacity of the knapsack.}
}
\value{
The maximum value that can be achieved within the given weight limit.
}
\description{
This function solves the knapsack problem using dynamic programming.
It optimally selects items to fit within a limited weight capacity.
}
\examples{
items <- data.frame(weight=c(1,2,3), value=c(6,10,12))
W <- 5
knapsack_dynamic(items, W)
}
